struct encode fix, simple version OK

+++ src/sei/pku/edu/cn/pattern/ConditionalPattern.java
@@ -2,7 +2,7 @@ package sei.pku.edu.cn.pattern;
 
 public class ConditionalPattern extends StatementPattern {
 
-	public ConditionalPattern(int varPos, int blockFlag) {
+	public ConditionalPattern(long varPos, int blockFlag) {
 		super(Label.CONDITIONAL, varPos, blockFlag);
 		
 	}
